requirements.txt
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
pydantic==2.9.2
passlib==1.7.4
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
pymongo==4.10.1
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
starlette==0.41.2
typer==0.12.5
ujson==5.10.0
bcrypt==4.2.0
uvicorn==0.32.0
watchfiles==0.24.0
PyJWT==1.7.1
confluent-kafka==2.6.1
python-jose==3.3.0
